refactor: Raise `DoesNotExist` error for non-existing emails
This update raises `User.DoesNotExist` for user emails that does not
exist instead of `ValidationError`. This ensures best practices and
consistency.

Also, this update adds a check to verify that all atributes are valid
and exists while creating or updating user objects. The implementation
is in the `set_extra_attributes` method.