User Model Cleanup & Refactoring (#6)
* chore: Update documentation with few styling

* chore: Update requirements.txt file with update packages

* feat: Register models in admin dashboard

* refactor: Update models to fix errors and handle minor validations

* feat: Add initial database model migration

* refactor: Raise `DoesNotExist` error for non-existing emails

This update raises `User.DoesNotExist` for user emails that does not
exist instead of `ValidationError`. This ensures best practices and
consistency.

Also, this update adds a check to verify that all atributes are valid
and exists while creating or updating user objects. The implementation
is in the `set_extra_attributes` method.

* chore: Fix typo when throwing `AttributeError` in `set_extra_attributes`